shuffle=None

+++ tets.py
@@ -53,7 +53,7 @@ def get_pretrained_model(include_top=False, pretrain_kind='imagenet'):
 class SiameseNetwork(nn.Module):
     def __init__(self, include_top=False):
         super(SiameseNetwork, self).__init__()
-        self.pretrained_model = get_pretrained_model(include_top, pretrain_kind='vggface2')
+        self.pretrained_model = get_pretrained_model(include_top, pretrain_kind='vggface2').half()
         self.ll1 = nn.Linear(4096, 100)
         self.relu = nn.ReLU()
         self.sigmod = nn.Sigmoid()
